Fix led_vu.h

+++ components/led_strip/led_vu.h
@@ -22,8 +22,8 @@ static struct led_strip_t* led_vu;
 void led_vu_progress_bar(int pct);
 void led_vu_display(int vu_l, int vu_r, bool comet);
 void led_vu_spin_dial(int rate, int gain, bool comet);
-void led_vu_spectrum(uint8_t* data, uint8_t offset, uint8_t length);
+void led_vu_spectrum(uint8_t* data, uint16_t offset, uint16_t length);
 void led_vu_color_all(uint8_t r, uint8_t g, uint8_t b);
-void led_vu_data(uint8_t* data, uint8_t offset, uint8_t length);
+void led_vu_data(uint8_t* data, uint16_t offset, uint16_t length);
 void led_vu_clear();
